requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==21.2.0
branca==0.4.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Fiona==1.8.20
Flask==2.0.2
flatbuffers==2.0
folium==0.12.1.post1
fonttools==4.28.2
gast==0.4.0
geopandas==0.10.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googlemaps==4.5.3
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
munch==2.5.0
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
osmnx==1.1.2
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pyproj==3.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
Rtree==0.9.7
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
Shapely==1.8.0
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==1.2.2
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
